+++ packages/app/src/request.ts
@@ -40,9 +40,9 @@ export const getHostname = (req: Request): string | undefined => {
   return index !== -1 ? host.substring(0, index) : host
 }
 
-export const getIP = (req: Request): string | undefined => proxyAddr(req, trustRemoteAddress(req)).replace(/^.*:/, '') // striping the redundant prefix addeded by OS to IPv4 address
+export const getIP = (req: Pick<IncomingMessage, 'headers' | 'connection'>): string | undefined => proxyAddr(req, trustRemoteAddress(req)).replace(/^.*:/, '') // striping the redundant prefix addeded by OS to IPv4 address
 
-export const getIPs = (req: Request): string[] | undefined => all(req, trustRemoteAddress(req))
+export const getIPs = (req: Pick<IncomingMessage, 'headers' | 'connection'>): string[] | undefined => all(req, trustRemoteAddress(req))
 
 export const getSubdomains = (req: Request, subdomainOffset = 2): string[] => {
   const hostname = getHostname(req)