+++ packages/proxy-addr/rollup.config.js
@@ -1,6 +1,7 @@
 import cfg from '../../build/cfgBuilder'
+import deps from '../../build/deps'
 import { dependencies } from './package.json'
 
 export default cfg({
-  external: dependencies,
+  external: deps(dependencies),
 })