+++ packages/router/rollup.config.js
@@ -1,6 +1,0 @@
-import cfg from '../../build/cfgBuilder'
-import { dependencies } from './package.json'
-
-export default cfg({
-  external: dependencies,
-})