+++ packages/forwarded/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tinyhttp/forwarded
 
+## 0.5.4
+
+### Patch Changes
+
+- make types more flexible
+
+## 0.5.3
+
+### Patch Changes
+
+- Make types more flexible
+
 ## 0.5.2
 
 ### Patch Changes